Create a ShaderProgram class
Create a class to encapsulate all the functionality of shaders in
OpenGL. Each individual shader will be in its own .glsl file. This
class reads each shader file, compiles them, links them, and binds
the resulting shader program. All the user has to do is pass in
the path of the shader files relative to $(SolutionDir).

Later, this class will be expanded to handle uniforms and textures.
It might also be expanded to handle other shader types besides
vertex and fragment shaders (which is all it can accept for now),
such as geometry shaders.