Create a Mesh class
Create a class to encapsulate all the functionality of vertex
buffers, vertex arrays, and index buffers. Each mesh contains only
one vertex buffer, but can be divided into many submeshes (since
one object can be made of multiple different materials or have many
different colors). Each submesh has its own vertex array, index
buffer, and shader.

Use the Mesh class by passing in the vertex buffer data, the size
(in bytes) of the data, and the layout of the data to the
constructor. Then for each submesh pass in an index buffer, an
index buffer count, and a shader. Finally, in the render loop
call .render() once to render the entire mesh.

Currently, this class trades control over the setup of vertex data
for simplicity of use. This may change in the future.