Create basic vertex and fragment shaders
Create a new directory $(SolutionDir)/res/shaders/ and create two
.glsl files (OpenGL shader language). These are test shaders. For
now, the vertex shader takes in a vec3 position and changes it to
a vec4, and the fragment shader outputs the color red as a vec4.

To use a shader, pass its file path to the constructor of a
ShaderProgram. Ex: "res/shaders/<shader name>.glsl". A
ShaderProgram must be given both a vertex and fragment shader.