Render a square in the window
Use the Mesh and ShaderProgram classes to setup and render a
square. Create a ShaderProgram by passing in the file paths
of the individual shaders. Create the vertex data as a float
array of positions in Normalized Device Coordinates. Finally,
add an index buffer and a shader specifying the vertices of
each triangle and how to draw them. If working correctly there
should be a red square in the middle of the window.