Add keyboard input handling
Add a query for the ESCAPE key so that when it is pressed, the
window will close and the render loop will end.

GLFW has functions that allow the user to query whether or not a
certain key is pressed down. This query is run every frame from a
function that is called inside the render loop. In the future,
functionality for keys such as the space bar and WASD will be
added in this function.