Add uniform functions to ShaderProgram
Create functions in the ShaderProgram class that allow the user
to add uniforms to their shaders. Add an unordered map as a
member variable to the class to act as a cache for the location
of each uniform to increase efficiency.

OpenGL has many separate functions for the different uniform types,
and these new functions in the ShaderProgram class act very
similarly. Currently there are only functions to add floats (1-4)
and a 4x4 matrix. More uniform functions may be added in the future.