Create a Camera class
Create a class to encapsulate all the functionality of a camera.
A camera has a position vector and the Euler angles yaw and pitch.
With these values, the forward, right, and up direction vectors
can be calculated and combined into the view matrix.

The Camera class used GLM vectors and matrices. The view matrix is
calculated using the glm::lookAt() function. The camera will be
translated with keyboard input and rotated with mouse input.