Add callback functions to Main
Add functions to main that are called automatically by GLFW when a
certain event takes place. The mouse callback is executed when the
user moves the mouse cursor, and the scroll callback is executed
when the user moves the mouse scroll wheel. These functions will be
used to call the input methods of the camera class to process mouse
input.