Add a camera object to Main
Create a camera object and give it an initial position. Since the
camera will be needed in the GLFW callback functions to handle
user input, set it as the window's user pointer. In the callback
functions and in processInput(), call the necessary camera member
functions to handle the user input.