Render a cube in the window
Change the vertex and index data to represent a cube, which has 8
vertices and 6 sides, each made up of 2 triangles. Use the GLM
library to create the model, view, and projection matrices. Use
the camera to get the view matrix and the FOV value for the
projection matrix. Pass these matrices to the shader as uniforms.

The user should be able to move around the window with WASD, look
around using the mouse, and zoom in and out using the scroll whell.