Capture the mouse cursor
Currently, mouse position input only works when the cursor is over
the window. This limits the user's range of visibility. Capture
the mouse cursor to keep it in the center of the window and make it
invisible.

The mouse cursor is invisible and not usable outside the window when
it is captured. In the future when a menu system is added, mouse
capture will be disabled in the menus and enabled during play.