Enable VSync
Call the method glfwSwapInterval() with 1 as the argument to tell
GLFW to wait for 1 monitor refresh before swapping buffers.

If the program is run without VSync, the render loop could run
hundreds or thousands of times per second, which is not
necessary. Enable VSync to tie the FPS to the refresh rate of
your monitor (usually 60).