Add stb_image library
Add stb_image to $(SolutionDir)/dependencies/includes/. This should
be an Include Directory. Use <stb_image/stb_image.h> to #include
this library into the project.

stb_image is a single-header library used to convert images into
arrays of data. It can be used with many different image formats.
It will be used in this project to convert textures (which are just
images) into data that OpenGL can use.