Create a Texture class
Create a class that encapsulates all the functionality of textures.
This class uses the stb_image library to convert an image file into
data that OpenGL can use, and it uses that data to create a 2D
texture.

To use, pass in the file path of the texture from $(SolutionDir)
and an int that acts as the texture's ID as arguments. Then the
texture can be added to the shader to act as a uniform.