Render a textured block
Render a textured Minecraft grass block using the texture sheet.
Each vertex now contains texture coordinates with offsets to find
the texture on the texture sheet. Expand the vertex buffer to 24
vertices instead of 8 because with textures the vertices are no
longer identical (they have different texture coordinates).

The vertex shader now takes in the texture coordinates and offsets,
calculates final texture coordinates, and sends this data to the
fragment shader. The fragment shader now takes in the texture
coordinates and the texture itself as a uniform and outputs the
pixel color based on these values.