Separate block info
Take all the hard-coded data and constants related to blocks,
vertex buffers, and index buffers and move them to a separate
header file. Also create a function that will allow the user to
query for vertex data based on the block type and the block face.
Place this function in its own source file.