Create a chunk class
Create a class that encapsulates all the properties of a chunk. A
chunk holds the position of its (relative) origin and an array that
contains the type of each block within the chunk as a number. Using
this data and the data in the BlockInfo header, the chunk can fill
a given float array with vertex data and a given unsigned int array
with index data.

Currently, a block has 24 vertices and each vertex is made of 5
floats, making each block 480 bytes of memory. And, so far, each
vertex only contains a position and texture coordinates. One
method to save memory has already been implemented in this class:
Don't include a block face in the vertex buffer if the adjacent
face is solid. More methods of saving memory will be added in the
future.