Detect OpenGL errors
Add a piece of code to the render loop to check for any errors that
might be thrown by OpenGL functions. The error code (an int) will
be outputted to the console. The type of error can be found by
consulting the OpenGL documentation. The location of the error is
also not specified, and will have to be tracked down.