Add a method to the Chunk class
Add a method that returns whether or not the chunk has been changed
since the last time the vertex and index data have been retrieved.
This will allow the user to know when they need to retrieve the
data again and will be more useful when mining and placing blocks
is implemented.