Make vertex data a single unsigned int
Combine all the vertex data into a single unsigned int. Each vertex
position, when relative to the chunk origin, is in the range [0-16]
(the plan now is to make chunks 16x16x16) which can be represented
with 15 bits (5 each x, y, z). Similarly, the texture coordinates
are in the range [0-16] on the texture sheet and only need 5 bits
each.

Currently, each vertex only needs 25 bits. This could be reduced to
20 if the chunks are made 15x15x15 and if only 15 rows and columns
on the texture sheet are used, but this unnecessary at the moment.