Remove index buffer and layout vector
Remove everything to do with index buffers. Since each vertex is
now only 1 32-bit integer, using an index buffer uses more memory
that not using one. With an index buffer, each block uses 24 ints
for the vertex data and 36 ints for the index data. Now, each block
uses only 36 ints for the vertex data. The data is now rendered
with glDrawArrays instead of glDrawElements.

The layout vector of the Mesh class was used to tell OpenGL the
layout of our vertex data. Since each vertex is now only 1 int,
this vector is not needed.