Render a grid of chunks to the window
Create many chunks with different positions in the world and place
them all in a list. In the render loop, call the render method of
each chunk.

All chunks are static and the world is a fixed size. In the future,
chunks that are too far away from the player will be deleted from
the list of chunks and closer chunks will be added to the list.