Put block data on the heap
Place the 3D blocks array in a struct and create that struct on the
heap so that all the block data goes on the heap. A struct is used
because it is very difficult to create a 3D array on the heap and
so that all the functions dealing with block data can be separated
into the struct.