Reduce allocated memory for vertices
Currently, enough space for every vertex is being allocated on the
GPU. This is very wasteful because most blocks in a chunk will
either be air or covered up by other solid blocks. Only allocating
space for around 35000 vertices per chunk (an arbitrary number that
will probably change later) is much more space efficient than
allocating space for the maximum number of vertices in a chunk
(2,359,296).

This solution is not the best since some chunks will need far less
than 35000 vertices while some will need more. In the future,
chunks will probably be static and, if changed, an entirely new
vertex buffer with exactly the space required will be generated.