Dynamically create new meshes
Whenever the state of a chunk changes, a new mesh and vertex buffer
will be created. This will allow all vertex buffers to be created
with the exact amount of memory required and will allow the vertex
buffers to be declared as static.

This is done to optimize for memory instead of time. It normally
takes less than 16 milliseconds (the time between each buffer swap
with vsync enabled) to create and fill a vertex buffer. Ideas for
the future include making only the chunks immediately around the
player dynamic and all other chunks static, or creating and filling
vertex buffers on a separate thread.