Redo the Chunk class with sections
Split each chunk into 16 sections. Each section is 16x16x16 blocks
and has its own mesh. Using sections will require only small
portions of a chunk to be recreated whenever there is an update.
Also, some sections may be entirely omitted if they are all air
(usually over half the sections in a chunk are all air). This will
save on memory since the block data in that section does not need
to be stored.

Chunk updates and omitting entire sections if they are all air
still needs to be implemented.