Allow chunks to see their neighbors
Give each chunk pointers to its 4 neighboring chunks. This will
allow chunks to retrieve the blocks of the neighboring chunks and
determine if it needs to render block faces along chunk borders.

Chunks need all their neighbors to find their vertex data, so this
step occurs after all chunks are created. The current world is a
fixed number of chunks and the chunks on the edge of the world draw
its outward faces by default.